Pagina_Cidepint/admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Daina Fernandez <day9311@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.3"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
flask-session = "^0.5.0"
flask-bcrypt = "^1.0.1"
flask-mail = "^0.9.1"
marshmallow = "^3.20.1"
pyjwt = "^2.8.0"
flask-cors = "<4.0"
flask-marshmallow = "<0.15"
authlib = "<1.4"
cryptography = "<39"
requests = "2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
flask-shell-ipython = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
